feat: move ToastProvider to base directory
